src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ../pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via web3
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.2
    # via reflex
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
appnope==0.1.4
    # via ipykernel
asttokens==2.4.1
    # via stack-data
attrs==24.2.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
bitarray==2.9.2
    # via eth-account
build==1.2.1
    # via reflex
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   reflex
    #   reflex-hosting-cli
    #   requests
ckzg==2.0.0
    # via eth-account
click==8.1.7
    # via
    #   typer
    #   uvicorn
comm==0.2.2
    # via ipykernel
cytoolz==0.12.3
    # via eth-utils
debugpy==1.8.5
    # via ipykernel
decorator==5.1.1
    # via ipython
dill==0.3.8
    # via reflex
dnspython==2.6.1
    # via pymongo
docutils==0.21.2
    # via readme-renderer
eth-abi==5.1.0
    # via
    #   eth-account
    #   web3
eth-account==0.13.3
    # via web3
eth-hash==0.7.0
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.5.1
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.1.0
    # via eth-account
eth-typing==5.0.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==5.0.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
executing==2.1.0
    # via stack-data
fastapi==0.110.3
    # via reflex
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gunicorn==22.0.0
    # via reflex
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
hexbytes==1.2.1
    # via
    #   eth-account
    #   eth-rlp
    #   eth-utils
    #   web3
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   defimoney-monitor (../pyproject.toml)
    #   python-telegram-bot
    #   reflex
    #   reflex-hosting-cli
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via twine
ipykernel==6.29.5
    # via defimoney-monitor (../pyproject.toml)
ipython==8.27.0
    # via ipykernel
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.0.2
    # via keyring
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   reflex
    #   starlette-admin
jupyter-client==8.6.2
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
keyring==25.3.0
    # via twine
lazy-loader==0.4
    # via reflex
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.4.0
    # via
    #   jaraco-classes
    #   jaraco-functools
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via ipykernel
nh3==0.2.18
    # via readme-renderer
packaging==24.1
    # via
    #   build
    #   gunicorn
    #   ipykernel
    #   lazy-loader
    #   reflex
parsimonious==0.10.0
    # via eth-abi
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pip==24.2
    # via pipdeptree
pipdeptree==2.16.2
    # via reflex-hosting-cli
pkginfo==1.10.0
    # via twine
platformdirs==4.2.2
    # via
    #   jupyter-core
    #   reflex
    #   reflex-hosting-cli
prompt-toolkit==3.0.47
    # via ipython
psutil==6.0.0
    # via
    #   ipykernel
    #   reflex
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycryptodome==3.20.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.8.2
    # via
    #   eth-account
    #   fastapi
    #   reflex
    #   reflex-hosting-cli
    #   sqlmodel
    #   web3
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via
    #   ipython
    #   readme-renderer
    #   rich
pymongo==4.8.0
    # via defimoney-monitor (../pyproject.toml)
pyproject-hooks==1.1.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   reflex-hosting-cli
python-dotenv==1.0.1
    # via defimoney-monitor (../pyproject.toml)
python-engineio==4.9.1
    # via
    #   python-socketio
    #   reflex
python-multipart==0.0.9
    # via
    #   reflex
    #   starlette-admin
python-socketio==5.11.3
    # via reflex
python-telegram-bot==21.5
    # via defimoney-monitor (../pyproject.toml)
pyunormalize==15.1.0
    # via web3
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
readme-renderer==44.0
    # via twine
redis==5.0.8
    # via reflex
reflex==0.5.10
    # via
    #   defimoney-monitor (../pyproject.toml)
    #   reflex-chakra
reflex-chakra==0.5.10
    # via reflex
reflex-hosting-cli==0.1.13
    # via reflex
regex==2024.7.24
    # via parsimonious
requests==2.32.3
    # via
    #   requests-toolbelt
    #   twine
    #   web3
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.8.0
    # via
    #   defimoney-monitor (../pyproject.toml)
    #   reflex
    #   reflex-hosting-cli
    #   twine
    #   typer
rlp==4.0.1
    # via
    #   eth-account
    #   eth-rlp
setuptools==70.1.1
    # via reflex
shellingham==1.5.4
    # via typer
simple-websocket==1.0.0
    # via python-engineio
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.32
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.21
    # via reflex
stack-data==0.6.3
    # via ipython
starlette==0.37.2
    # via
    #   fastapi
    #   starlette-admin
starlette-admin==0.14.1
    # via reflex
tabulate==0.9.0
    # via reflex-hosting-cli
tomlkit==0.13.2
    # via reflex
toolz==0.12.1
    # via cytoolz
tornado==6.4.1
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
twine==5.1.1
    # via reflex
typer==0.12.5
    # via
    #   reflex
    #   reflex-hosting-cli
types-requests==2.32.0.20240712
    # via web3
typing-extensions==4.12.2
    # via
    #   alembic
    #   eth-typing
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   web3
urllib3==2.2.2
    # via
    #   requests
    #   twine
    #   types-requests
uvicorn==0.30.6
    # via reflex
watchdog==4.0.2
    # via reflex
watchfiles==0.24.0
    # via reflex
wcwidth==0.2.13
    # via prompt-toolkit
web3==7.2.0
    # via defimoney-monitor (../pyproject.toml)
websockets==13.0.1
    # via
    #   reflex-hosting-cli
    #   web3
wheel==0.44.0
    # via reflex
wrapt==1.16.0
    # via reflex
wsproto==1.2.0
    # via simple-websocket
yarl==1.9.4
    # via aiohttp
zipp==3.20.1
    # via importlib-metadata
